requirements.txt
# VLM Project with Qwen2.5-VL for Video Reasoning
# Video + Text Query → Chain-of-Thought Reasoning

torch>=2.0.0
transformers>=4.37.0
accelerate>=0.20.0
decord>=0.6.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
pyyaml>=6.0
einops>=0.7.0
sentencepiece>=0.1.99
protobuf>=3.20.0
# For video processing
av>=10.0.0
# For dataset handling
datasets>=2.14.0
# For logging
wandb>=0.15.0  # optional
# For API calls (if using OpenAI API)
openai>=1.0.0  # optional
